disable memory file, put up jason's code to fix negative problem, add fuse preprocess param, and make tpu-mlir an option
