+++ sophgo_mq/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '1.0.4'
+__version__ = '1.0.6'
