+++ sophgo_mq/mlir/__init__.py
@@ -1,0 +1,1 @@
+from .tpu_utils import *
