+++ sophgo_mq/prepare_by_platform.py
@@ -11,8 +11,9 @@ from torch.fx import Tracer
 from torch.fx.graph_module import GraphModule
 from torch.quantization.quantize_fx import _swap_ff_with_fxff
 from torch.quantization import QConfig
-from sophgo_mq.convert_deploy import convert_deploy, param_to_idx_dict, convert_merge_bn, deepcopy_graphmodule
+from sophgo_mq.convert_deploy import convert_deploy, convert_merge_bn, deepcopy_graphmodule
 from sophgo_mq.utils.utils import generate_random_string
+from sophgo_mq.mlir.tpu_utils import param_to_idx_dict
 
 from sophgo_mq.fake_quantize import (
     LearnableFakeQuantize,
@@ -566,12 +567,34 @@ def prepare_constant_dict(graph: torch.fx.Graph, model: torch.nn.Module):
             constant_dict[node.target] = _get_attrs(model, node.target)
     return constant_dict
 
+def prepare_dynamic_param_update(prepared: GraphModule, cuda: int = None, input_shape_dict: list = None, chip: str = "CV181X",  model_name: str = "sophgo_mq_qmodel", net_type: str = "CNN"):
+    print('>>>Gets the mapping between the pytorch param parameter and the parameter to be updated in the tpu-mlir...')
+    module_tmp = deepcopy_graphmodule(prepared)
+    module_tmp = module_tmp.cuda(cuda)
+    module_tmp.eval()
+    images = torch.zeros(*list(input_shape_dict.values())[0], device='cuda')
+    module_tmp(images)
+
+    idx = 100
+    convert_merge_bn(module_tmp.eval())
+    global param_to_idx_dict
+    for name, param in module_tmp.named_parameters():
+        param = param.view(-1)
+        old_data0 = copy.deepcopy(param.data[0])
+        param.data[0] = idx
+        unique_id = generate_random_string(30)
+        print(f'prepare_by_platform, name: {name}, data0:{idx}, unique_id: {unique_id}, old_data0:{old_data0}')
+        param_to_idx_dict[name] = [idx,unique_id]
+        idx = idx + 1
+    convert_deploy(module_tmp, net_type, input_shape_dict=input_shape_dict,
+        model_name='{}'.format(model_name), not_gen_bmodel = True,
+        output_path='/tmp/', bf16_mix_prec = False, mlir_deploy=False, chip=chip)
+        #output_path='/dev/shm/', bf16_mix_prec = False, mlir_deploy=False, chip=chip)
+    print('>>>Gets the mapping, end')
+
 def prepare_by_platform(
-        arch: str,
         model: torch.nn.Module,
-        net_type = 'CNN',
         input_shape_dict: list = None, cuda:int = None,
-        val_loader = None,
         prepare_custom_config_dict: Dict[str, Any] = {},
         custom_tracer: Tracer = None):
     """
@@ -600,6 +623,10 @@ def prepare_by_platform(
     chip=quant_dict['chip']
     strategy=quant_dict['strategy']
     not_dynamic_update_param = False if 'not_dynamic_update_param' not in quant_dict else quant_dict['not_dynamic_update_param']
+    use_tpu_mlir_fwd = False if 'use_tpu_mlir_fwd' not in quant_dict else quant_dict['use_tpu_mlir_fwd']
+    net_type = 'CNN' if 'strategy' not in quant_dict else quant_dict['strategy']
+    model_name = "sophgo_mq_qmodel" if 'model_name' not in quant_dict else quant_dict['model_name']
+
     logger.info("Quantize model Scheme: {} Mode: {}".format(quant_dict['strategy'], model_mode))
     qconfig = get_qconfig_by_platform(quant_dict, extra_qconfig_dict)
 
@@ -665,27 +692,6 @@ def prepare_by_platform(
                 if inspect.ismethod(_attr):
                     _attr = types.MethodType(getattr(_type, attr_name), cur_module)
                 setattr(cur_module, attr_name, _attr)
-    if not not_dynamic_update_param:
-        print('>>>Gets the mapping between the pytorch param parameter and the parameter to be updated in the tpu-mlir...')
-        module_tmp = deepcopy_graphmodule(prepared)
-        module_tmp = module_tmp.cuda(cuda)
-        module_tmp.eval()
-        images = torch.zeros(*list(input_shape_dict.values())[0], device='cuda')
-        module_tmp(images)
-        
-        idx = 100
-        convert_merge_bn(module_tmp.eval())
-        global param_to_idx_dict
-        for name, param in module_tmp.named_parameters():
-            param = param.view(-1)
-            old_data0 = copy.deepcopy(param.data[0])
-            param.data[0] = idx
-            unique_id = generate_random_string(30)
-            print(f'prepare_by_platform, name: {name}, data0:{idx}, unique_id: {unique_id}, old_data0:{old_data0}')
-            param_to_idx_dict[name] = [idx,unique_id]
-            idx = idx + 1
-        convert_deploy(module_tmp, net_type, input_shape_dict=input_shape_dict,
-            model_name='{}'.format(arch), not_gen_bmodel = True,
-            output_path='/dev/shm/', bf16_mix_prec = False, mlir_deploy=False, chip=chip)
-        print('>>>Gets the mapping, end')
+    if (not not_dynamic_update_param) and use_tpu_mlir_fwd:
+        prepare_dynamic_param_update(prepared, cuda, input_shape_dict, chip, model_name, net_type)
     return prepared
